Split PersonConditionTextures into separated document
